@@ -18,9 +18,9 @@
     "typescript:transpile:dev": "echo 'Use `yarn docs:typescript'` instead && exit 1"
   },
   "dependencies": {
-    "@babel/core": "^7.19.0",
+    "@babel/core": "^7.19.1",
     "@babel/plugin-transform-object-assign": "^7.18.6",
-    "@babel/runtime-corejs2": "^7.19.0",
+    "@babel/runtime-corejs2": "^7.19.2",
     "@date-io/core": "^2.15.0",
     "@date-io/date-fns-jalali": "^2.15.0",
     "@docsearch/react": "^3.2.1",
@@ -31,23 +31,23 @@
     "@fortawesome/fontawesome-svg-core": "^6.2.0",
     "@fortawesome/free-solid-svg-icons": "^6.2.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "@mui/base": "5.0.0-alpha.98",
+    "@mui/base": "5.0.0-alpha.99",
     "@mui/docs": "^5.10.6",
     "@mui/icons-material": "^5.10.6",
-    "@mui/joy": "5.0.0-alpha.46",
-    "@mui/lab": "5.0.0-alpha.100",
-    "@mui/material": "^5.10.6",
-    "@mui/material-next": "6.0.0-alpha.54",
-    "@mui/styled-engine": "^5.10.6",
+    "@mui/joy": "5.0.0-alpha.47",
+    "@mui/lab": "5.0.0-alpha.101",
+    "@mui/material": "^5.10.7",
+    "@mui/material-next": "6.0.0-alpha.55",
+    "@mui/styled-engine": "^5.10.7",
     "@mui/styled-engine-sc": "^5.10.6",
-    "@mui/styles": "^5.10.6",
-    "@mui/system": "^5.10.6",
+    "@mui/styles": "^5.10.7",
+    "@mui/system": "^5.10.7",
     "@mui/types": "^7.2.0",
-    "@mui/x-data-grid": "^5.17.2",
-    "@mui/x-data-grid-generator": "^5.17.2",
-    "@mui/x-data-grid-pro": "^5.17.2",
-    "@mui/x-date-pickers": "^5.0.1",
-    "@mui/x-date-pickers-pro": "^5.0.1",
+    "@mui/x-data-grid": "^5.17.4",
+    "@mui/x-data-grid-generator": "^5.17.4",
+    "@mui/x-data-grid-pro": "^5.17.4",
+    "@mui/x-date-pickers": "^5.0.3",
+    "@mui/x-date-pickers-pro": "^5.0.3",
     "@mui/x-license-pro": "^5.17.0",
     "@trendmicro/react-interpolate": "^0.5.5",
     "@types/autosuggest-highlight": "^3.2.0",
@@ -56,13 +56,13 @@
     "@types/markdown-to-jsx": "^7.0.1",
     "@types/react-dom": "^18.0.6",
     "@types/react-swipeable-views": "^0.13.1",
-    "@types/react-swipeable-views-utils": "^0.13.3",
+    "@types/react-swipeable-views-utils": "^0.13.4",
     "@types/react-virtualized": "^9.21.21",
     "@types/react-window": "^1.8.5",
     "@types/styled-components": "5.1.26",
     "accept-language": "^3.0.18",
     "ast-types": "^0.14.2",
-    "autoprefixer": "^10.4.11",
+    "autoprefixer": "^10.4.12",
     "autosuggest-highlight": "^3.3.4",
     "babel-plugin-module-resolver": "^4.1.0",
     "babel-plugin-optimize-clsx": "^2.6.2",
@@ -92,7 +92,7 @@
     "lz-string": "^1.4.4",
     "markdown-to-jsx": "^7.1.7",
     "material-ui-popup-state": "^2.0.1",
-    "next": "12.2.5",
+    "next": "12.3.1",
     "notistack": "2.0.5",
     "nprogress": "^0.2.0",
     "postcss": "^8.4.16",
@@ -102,11 +102,11 @@
     "react-dom": "^18.2.0",
     "react-draggable": "^4.4.5",
     "react-final-form": "^6.5.9",
-    "react-imask": "^6.4.2",
+    "react-imask": "^6.4.3",
     "react-intersection-observer": "^9.4.0",
     "react-is": "^18.2.0",
     "react-number-format": "^4.9.4",
-    "react-router-dom": "^6.4.0",
+    "react-router-dom": "^6.4.1",
     "react-spring": "^8.0.27",
     "react-swipeable-views": "^0.14.0",
     "react-transition-group": "^4.4.5",
