@@ -197,7 +197,7 @@ describe('replaceMarkdownLinks', () => {
       [ButtonUnstyled](/api/button-unstyled)
       [TabPanelUnstyled](/api/tab-panel-unstyled)
       [TabsListUnstyled](/api/tabs-list-unstyled)
-      [TrapFocus](/api/trap-focus)
+      [FocusTrap](/api/focus-trap)
       [ClickAwayListener](/api/click-away-listener)
       [IconButton](/api/icon-button)
       [LoadingButton](/api/loading-button)
@@ -216,7 +216,7 @@ describe('replaceMarkdownLinks', () => {
       [ButtonUnstyled](/base/api/button-unstyled)
       [TabPanelUnstyled](/base/api/tab-panel-unstyled)
       [TabsListUnstyled](/base/api/tabs-list-unstyled)
-      [TrapFocus](/base/api/trap-focus)
+      [FocusTrap](/base/api/focus-trap)
       [ClickAwayListener](/base/api/click-away-listener)
       [IconButton](/material-ui/api/icon-button)
       [LoadingButton](/material-ui/api/loading-button)
@@ -365,7 +365,7 @@ describe('replaceMarkdownLinks', () => {
 
   it('does not trasform material-ui', () => {
     expect(
-      replaceMarkdownLinks(`[Trap Focus](https://mui.com/material-ui/react-trap-focus/)`),
-    ).to.equal(`[Trap Focus](https://mui.com/material-ui/react-trap-focus/)`);
+      replaceMarkdownLinks(`[Focus Trap](https://mui.com/material-ui/react-trap-focus/)`),
+    ).to.equal(`[Focus Trap](https://mui.com/material-ui/react-trap-focus/)`);
   });
 });
