@@ -138,9 +138,9 @@ export default function useListbox<TOption>(props: UseListboxParameters<TOption>
       });
     };
 
-  const createHandleOptionMouseOver =
+  const createHandleOptionPointerOver =
     (option: TOption, other: Record<string, React.EventHandler<any>>) =>
-    (event: React.MouseEvent) => {
+    (event: React.PointerEvent) => {
       other.onMouseOver?.(event);
       if (event.defaultPrevented) {
         return;
@@ -298,11 +298,11 @@ export default function useListbox<TOption>(props: UseListboxParameters<TOption>
       ...otherHandlers,
       'aria-disabled': optionState.disabled || undefined,
       'aria-selected': optionState.selected,
-      tabIndex: getOptionTabIndex(optionState),
       id: optionIdGenerator(option, index),
       onClick: createHandleOptionClick(option, otherHandlers),
-      onMouseOver: createHandleOptionMouseOver(option, otherHandlers),
+      onPointerOver: createHandleOptionPointerOver(option, otherHandlers),
       role: 'option',
+      tabIndex: getOptionTabIndex(optionState),
     };
   };
 
