@@ -1,5 +1,6 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
+import clsx from 'clsx';
 import { unstable_composeClasses as composeClasses } from '@mui/base';
 import styled, { rootShouldForwardProp } from '../styles/styled';
 import useThemeProps from '../styles/useThemeProps';
@@ -30,7 +31,7 @@ const DialogContentTextRoot = styled(Typography, {
 
 const DialogContentText = React.forwardRef(function DialogContentText(inProps, ref) {
   const props = useThemeProps({ props: inProps, name: 'MuiDialogContentText' });
-  const { children, ...ownerState } = props;
+  const { children, className, ...ownerState } = props;
   const classes = useUtilityClasses(ownerState);
 
   return (
@@ -40,6 +41,7 @@ const DialogContentText = React.forwardRef(function DialogContentText(inProps, r
       color="text.secondary"
       ref={ref}
       ownerState={ownerState}
+      className={clsx(classes.root, className)}
       {...props}
       classes={classes}
     />
@@ -59,6 +61,10 @@ DialogContentText.propTypes /* remove-proptypes */ = {
    * Override or extend the styles applied to the component.
    */
   classes: PropTypes.object,
+  /**
+   * @ignore
+   */
+  className: PropTypes.string,
   /**
    * The system prop that allows defining system overrides as well as additional CSS styles.
    */
