@@ -81,24 +81,30 @@ describe('<Slider />', () => {
     const { container } = render(
       <Slider onChange={handleChange} onChangeCommitted={handleChangeCommitted} value={0} />,
     );
+    stub(container.firstChild, 'getBoundingClientRect').callsFake(() => ({
+      width: 100,
+      height: 10,
+      bottom: 10,
+      left: 0,
+    }));
 
-    fireEvent.touchStart(container.firstChild, createTouches([{ identifier: 1 }]));
+    fireEvent.touchStart(container.firstChild, createTouches([{ identifier: 1, clientX: 0 }]));
     expect(handleChange.callCount).to.equal(1);
     expect(handleChangeCommitted.callCount).to.equal(0);
 
-    fireEvent.touchEnd(document.body, createTouches([{ identifier: 2 }]));
+    fireEvent.touchStart(document.body, createTouches([{ identifier: 2, clientX: 40 }]));
     expect(handleChange.callCount).to.equal(1);
     expect(handleChangeCommitted.callCount).to.equal(0);
 
-    fireEvent.touchMove(document.body, createTouches([{ identifier: 1 }]));
+    fireEvent.touchMove(document.body, createTouches([{ identifier: 1, clientX: 1 }]));
     expect(handleChange.callCount).to.equal(2);
     expect(handleChangeCommitted.callCount).to.equal(0);
 
-    fireEvent.touchMove(document.body, createTouches([{ identifier: 2 }]));
+    fireEvent.touchMove(document.body, createTouches([{ identifier: 2, clientX: 41 }]));
     expect(handleChange.callCount).to.equal(2);
     expect(handleChangeCommitted.callCount).to.equal(0);
 
-    fireEvent.touchEnd(document.body, createTouches([{ identifier: 1 }]));
+    fireEvent.touchEnd(document.body, createTouches([{ identifier: 1, clientX: 2 }]));
     expect(handleChange.callCount).to.equal(2);
     expect(handleChangeCommitted.callCount).to.equal(1);
   });
@@ -131,6 +137,34 @@ describe('<Slider />', () => {
     });
     // The mouse's button was released, stop the dragging session.
     expect(handleChange.callCount).to.equal(2);
+  });
+
+  it('should only fire onChange when the value changes', () => {
+    const handleChange = spy();
+    const { container } = render(<Slider defaultValue={20} onChange={handleChange} />);
+    stub(container.firstChild, 'getBoundingClientRect').callsFake(() => ({
+      width: 100,
+      left: 0,
+    }));
+
+    fireEvent.mouseDown(container.firstChild, {
+      buttons: 1,
+      clientX: 21,
+    });
+
+    fireEvent.mouseMove(document.body, {
+      buttons: 1,
+      clientX: 22,
+    });
+    // Sometimes another event with the same position is fired by the browser.
+    fireEvent.mouseMove(document.body, {
+      buttons: 1,
+      clientX: 22,
+    });
+
+    expect(handleChange.callCount).to.equal(2);
+    expect(handleChange.args[0][1]).to.deep.equal(21);
+    expect(handleChange.args[1][1]).to.deep.equal(22);
   });
 
   describe('prop: orientation', () => {
@@ -232,7 +266,7 @@ describe('<Slider />', () => {
       expect(slider).toHaveFocus();
     });
 
-    it('should support mouse events', () => {
+    it('should support touch events', () => {
       const handleChange = spy();
       const { container } = render(<Slider defaultValue={[20, 30]} onChange={handleChange} />);
       stub(container.firstChild, 'getBoundingClientRect').callsFake(() => ({
@@ -253,13 +287,14 @@ describe('<Slider />', () => {
       );
       fireEvent.touchMove(
         document.body,
-        createTouches([{ identifier: 1, clientX: 22, clientY: 0 }]),
+        createTouches([{ identifier: 1, clientX: 22.1, clientY: 0 }]),
       );
 
       expect(handleChange.callCount).to.equal(3);
       expect(handleChange.args[0][1]).to.deep.equal([21, 30]);
       expect(handleChange.args[1][1]).to.deep.equal([22, 30]);
-      expect(handleChange.args[2][1]).not.to.equal(handleChange.args[1][1]);
+      // TODO, consider not firing this change event since the values are the same to improve the DX.
+      expect(handleChange.args[2][1]).to.deep.equal([22, 30]);
     });
 
     it('should not react to right clicks', () => {
