@@ -1,2 +1,2 @@
-export { default } from '@mui/base/TrapFocus';
-export * from '@mui/base/TrapFocus';
+export { default } from '@mui/base/FocusTrap';
+export * from '@mui/base/FocusTrap';
